% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dic_calculations.R
\name{dic.calculations}
\alias{dic.calculations}
\title{Compute Deviance Information Criterion (DIC)}
\usage{
dic.calculations(X, Y, z, gammas, betas, phieps, Us, deviances, T1, T2)
}
\arguments{
\item{X}{Spline basis matrix (T x J)}

\item{Y}{Matrix of model outputs (T x m)}

\item{z}{Observational vector of length (T2 - T1 + 1)}

\item{gammas}{Array of gamma samples (nsim x J x H)}

\item{betas}{Array of beta samples (nsim x M x H+1)}

\item{phieps}{Matrix of precision samples (nsim x M)}

\item{Us}{Array of latent factor samples (nsim x T x H)}

\item{deviances}{Vector of deviance values (length nsim)}

\item{T1}{Starting time index for observational window}

\item{T2}{Ending time index for observational window}
}
\value{
A list containing DIC, effective number of parameters, mean deviance,
and deviance at posterior means
}
\description{
Calculates the Deviance Information Criterion and its components using posterior
samples from a Bayesian model fit.
Called by BM3 function during the MCMC sampling process.
}
