% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_posterior}
\alias{plot_posterior}
\title{Plot Posterior Distribution}
\usage{
plot_posterior(results, H, data = NULL, title = NULL)
}
\arguments{
\item{results}{List from \code{consolidate_results()}.}

\item{H}{Integer. Which H value to plot.}

\item{data}{Optional. Observations to overlay (default NULL).}

\item{title}{Optional plot title (default NULL).}
}
\value{
ggplot2 object.
}
\description{
Create a plot of posterior means and credible intervals.
}
\examples{
\dontrun{
plot_posterior(results, H = 2, title = "My Title")
}

}
