% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_processing.R
\name{posterior_mean_curve}
\alias{posterior_mean_curve}
\title{Compute Posterior Mean Curve}
\usage{
posterior_mean_curve(BFACT)
}
\arguments{
\item{BFACT}{A BFACT object with a $ysim element.}
}
\value{
Numeric vector of posterior means.
}
\description{
Returns the posterior mean curve from a BFACT object.
}
