% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_processing.R
\name{sample_posterior}
\alias{sample_posterior}
\title{Sampled Posterior Curves}
\usage{
sample_posterior(BFACT, ncurves = NULL)
}
\arguments{
\item{BFACT}{A BFACT object with a $ysim element.}
}
\value{
Numeric matrix of posterior curves.
}
\description{
Returns a sample of posterior curves from a BFACT object.
}
